fix(travis): fix broken deployments by ensuring GH_TOKEN repo url
This fixes the broken travis deployment scripts - which would always fail because npm run pages
would deploy to an http url which it had no access to. This specifies the repository option which
includes a GH_TOKEN so that travis can push. It also reorders the script to ensure semantic-release
is always run first, as this is the higher priority of the two publish scripts